+++ fox.common/src/main/java/com/minutch/fox/param/decoration/GoodsParam.java
@@ -15,7 +15,7 @@ public class GoodsParam {
     private String goodsModel;
     private BigDecimal goodsPrice;
     private BigDecimal inGoodsPrice;
-    private Integer stockNum;
+    private BigDecimal stockNum;
     private Integer stockWarn;
     private Long whId;
     private String whName;
