+++ fox.common/src/main/java/com/minutch/fox/param/decoration/OrderParam.java
@@ -22,7 +22,7 @@ public class OrderParam {
     private Long goodsId;
     private String goodsName;
     private String goodsModel;
-    private Integer goodsNum;
+    private BigDecimal goodsNum;
     private BigDecimal goodsPrice;
     private BigDecimal inGoodsPrice;
     private BigDecimal orderAmount;
