+++ fox.dao/src/main/java/com/minutch/fox/entity/decoration/StockDetail.java
@@ -6,12 +6,14 @@ import com.minutch.fox.entity.base.BaseEntity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode(callSuper = false)
 @Data
 public class StockDetail extends BaseEntity {
 
   private Long goodsId;
-  private Integer stockNum;
+  private BigDecimal stockNum;
   private String objType;
   private Long objId;
   private Long storeId;
