+++ fox.wap/src/main/java/com/minutch/fox/result/decoration/GoodsVO.java
@@ -19,7 +19,7 @@ public class GoodsVO {
     private String goodsModel;
     private BigDecimal goodsPrice;
     private BigDecimal inGoodsPrice;
-    private Integer stockNum;
+    private BigDecimal stockNum;
     private Integer stockWarn;
     private Long whId;
     private String whName;
