+++ fox.wap/src/main/java/com/minutch/fox/result/decoration/OrderVO.java
@@ -26,8 +26,8 @@ public class OrderVO {
     private Long goodsId;
     private String goodsName;
     private String goodsModel;
-    private Integer goodsNum;
-    private Integer useNum;
+    private BigDecimal goodsNum;
+    private BigDecimal useNum;
     private Integer handleStock;
     private BigDecimal goodsPrice;
     private BigDecimal inGoodsPrice;
